backend/requirements.txt
# Core dependencies
fastapi>=0.100.0
uvicorn>=0.23.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# Validation
pydantic>=2.0.0
email-validator>=2.0.0

# Configuration
python-dotenv>=1.0.0

# Image processing
Pillow>=10.0.0
python-multipart>=0.0.6
